Fix for test failures (events_logs_tests) due to loss of precision for timeouts. This was affecting timeout values passed to the SLEEP(duration) function, a 1.5 seconds timeout would be converted to a 1 second timeout.

+++ sql/item_func.cc
@@ -3473,7 +3473,7 @@ static int interruptible_wait(THD *thd, pthread_cond_t *cond,
 {
   int error;
   struct timespec abstime;
-  ulonglong slice, timeout= (ulonglong) time * ULL(1000000000);
+  ulonglong slice, timeout= (ulonglong) (time * 1000000000.0);
 
   do
   {
